requirements.txt
Django==1.10.7
django-markdownx==2.0.21
django-widget-tweaks==1.4.1
Markdown==2.6.8
olefile==0.44
Pillow==4.1.1
